requirements.txt
flask~=2.0.2
flask_httpauth~=4.4.0
gunicorn~=20.1.0
scikit-learn~=1.0
Pillow~=8.3.2
spacy~=3.1.3
setuptools~=57.0.0
tqdm~=4.62.3
numpy~=1.19.5
tensorflow~=2.6.0
matplotlib~=3.4.3